fastapi_backend/requirements.txt
fastapi>=0.110.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.30.0
python-multipart>=0.0.9,<0.1.0
aiofiles>=23.2.1,<24.0.0
pydantic>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Supabase
supabase>=2.8.0,<3.0.0

# AWS SDK
boto3>=1.34.34,<1.35.0
botocore>=1.34.34,<1.35.0

# Google AI/Gemini - Updated to be compatible with langchain-google-genai
google-generativeai>=0.8.0,<0.9.0
aiohttp>=3.9.1,<4.0.0

# YouTube transcript
youtube-transcript-api>=0.6.2,<0.7.0

# Additional dependencies
python-json-logger>=2.0.7,<3.0.0
requests>=2.31.0,<3.0.0
typing-extensions>=4.12.2,<5.0.0 

yt-dlp>=2023.12.30
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# SSL and certificate handling
certifi>=2023.0.0
urllib3>=2.0.0
cryptography>=41.0.0

# HTTP client with better SSL support  
httpx>=0.25.0 